+++ .env-template
@@ -1,0 +1,2 @@
+token=
+#paste the token above without "" around it
