+++ commands/automod/antispam.js
@@ -56,7 +56,7 @@ module.exports.run = async (client, message, args) => {
 module.exports.config = {
 	name: "antispam",
 	aliases: ["as"],
-	category: "Automod",
+	
 	usage: "*antispam 'on or off'",
 	description: "Turns antispam on or off"
 }
