+++ commands/misc/ping.js
@@ -25,7 +25,7 @@ module.exports.run = async(client, message, args) =>  {
    module.exports.config = {
     name: "ping",
     aliases: ["p"],
-    category: "Misc",
+    
     usage: "*ping",
     description: "Sends the bot's ping and latency"
 };
