+++ utils/utilsMain.js
@@ -1,7 +1,7 @@
 const { Client } = require("discord.js");
 
 const moment = require("moment");
-
+require('colors')
 /**
  * Big boi collection of all commands used within the client globally, Usage: client.loadCommands(...)
  * @param {Client} client
@@ -23,4 +23,33 @@ module.exports = (client) => {
         console.log(`[${moment().format('D/M/Y HH:mm:ss.SSS').bold.blue}] [${type.green}] [${title.yellow}] ${msg}`);
     };
 
+    let lastCategoryLoaded;
+    client.loadCommand = (category, commandName, dontLog) => {
+    try {
+        const props = require(`${process.cwd()}/commands/${category}/${commandName}`);
+        if(lastCategoryLoaded !== category){
+        client.log("Load", `Starting to load all commands from the category ${category}`);
+        lastCategoryLoaded = category;
+        }
+        if (!dontLog) {
+        client.log("Load", `  ${"=>".blue} ${"Loading Command:".white} ${props.config.name.green}.`);
+        }
+        if (props.init) {
+        props.init(client);
+        }
+        if (category) props.category = category
+        client.commands.set(props.config.name, props);
+        props.config.aliases.forEach(alias => {
+        client.aliases.set(alias, props.config.aliases);
+        });
+        return {
+        res: true
+        };
+    } catch (e) {
+        console.log(e)
+        return {
+        err: `Unable to load command ${commandName} in ${category}: ${e}`
+        };
+    }
+};
 }
